ops-agent/pyproject.toml
[project]
name = "ops-agent"
version = "0.1.0"
description = "Standalone OpsCopilot agent service powered by Google ADK"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "httpx>=0.27",
  "psycopg[binary]>=3.2",
  "google-adk[eval]>=1.0",
]

[tool.uv]
managed = true
